Fixes following comments from JeanMertz and lukesteensen
